refactor: break type mappings up in multiple files
